third_party/blink/web_tests/pointer-lock/pointerlockchange-event-on-lock-lost-expected.txt

Test that pointerlockchange event is dispatched when lock is lost.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.testRunner is defined.
     Lock targetDiv1.
PASS onpointerlockchange received after: Lock targetDiv1.
PASS document.pointerLockElement is targetDiv1
     Host initiated unlock.
PASS onpointerlockchange received after: Host initiated unlock.
PASS document.pointerLockElement is null
PASS successfullyParsed is true

TEST COMPLETE
